### 2016-04-13 6742b4a
Option to remove stopwords from word tokenizer
