+++ tests/testthat/test-utils.R
@@ -9,3 +9,8 @@ test_that("Inputs are verified correct", {
   expect_error(check_input(list(a = "a", b = letters)))
   expect_error(check_input(list(a = "a", b = 2)))
 })
+
+test_that("Stopwords are removed", {
+  expect_equal(remove_stopwords(letters[1:5], stopwords = c("d", "e")),
+               letters[1:3])
+})
