### 2016-04-13 7e03807
Remove stopwords from word stem tokenizer
