+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: tokenizers
 Type: Package
 Title: Tokenize Text
-Version: 0.1.1
+Version: 0.1.1.9000
 Date: 2016-04-03
 Description: Convert natural language text into tokens. The tokenizers have a
     consistent interface and are compatible with Unicode, thanks to being built
