### 2016-04-13 3172413
Add Lucene/Solr stopwords
Fixes #9
