+++ data-raw/lucene-stopwords/stopwords_da.txt
@@ -1,0 +1,106 @@
+# | From svn.tartarus.org/snowball/trunk/website/algorithms/danish/stop.txt
+# | This file is distributed under the BSD License.
+# | See http://snowball.tartarus.org/license.php
+# | Also see http://www.opensource.org/licenses/bsd-license.html
+# |  - Encoding was converted to UTF-8.
+# |  - This notice was added.# 
+# | A Danish stop word list. Comments begin with vertical bar. Each stop
+# | word is at the start of a line.# 
+# | This is a ranked list (commonest to rarest) of stopwords derived from
+# | a large text sample.
+
+
+og
+i
+jeg
+det
+at
+en
+den
+til
+er
+som
+på
+de
+med
+han
+af
+for
+ikke
+der
+var
+mig
+sig
+men
+et
+har
+om
+vi
+min
+havde
+ham
+hun
+nu
+over
+da
+fra
+du
+ud
+sin
+dem
+os
+op
+man
+hans
+hvor
+eller
+hvad
+skal
+selv
+her
+alle
+vil
+blev
+kunne
+ind
+når
+være
+dog
+noget
+ville
+jo
+deres
+efter
+ned
+skulle
+denne
+end
+dette
+mit
+også
+under
+have
+dig
+anden
+hende
+mine
+alt
+meget
+sit
+sine
+vor
+mod
+disse
+hvis
+din
+nogle
+hos
+blive
+mange
+ad
+bliver
+hendes
+været
+thi
+jer
+sådan
