+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+export(stopwords)
 export(tokenize_characters)
 export(tokenize_lines)
 export(tokenize_ngrams)

[…]
