### 2016-04-13 2add653
Fix documentation/description for stopwords
