+++ data-raw/lucene-stopwords/stopwords_ru.txt
@@ -1,0 +1,223 @@
+# | From svn.tartarus.org/snowball/trunk/website/algorithms/russian/stop.txt
+# | with some cleaning from @dselivanov
+# | This file is distributed under the BSD License.
+# | See http://snowball.tartarus.org/license.php
+# | Also see http://www.opensource.org/licenses/bsd-license.html
+# |  - Encoding was converted to UTF-8.
+# |  - This notice was added.
+# | a russian stop word list. comments begin with vertical bar. each stop
+# | word is at the start of a line.
+# | this is a ranked list (commonest to rarest) of stopwords derived from
+# | a large text sample.
+# | letter `ё' is translated to `е'.
+
+и
+в
+во
+не
+что
+он
+на
+я
+с
+со
+как
+а
+то
+все
+она
+так
+его
+но
+да
+ты
+к
+у
+же
+вы
+за
+бы
+по
+только
+ее
+мне
+было
+вот
+от
+меня
+еще
+нет
+о
+из
+ему
+теперь
+когда
+даже
+ну
+вдруг
+ли
+если
+уже
+или
+ни
+быть
+был
+него
+до
+вас
+нибудь
+опять
+уж
+вам
+сказал
+ведь
+там
+потом
+себя
+ничего
+ей
+может
+они
+тут
+где
+есть
+надо
+ней
+для
+мы
+тебя
+их
+чем
+была
+сам
+чтоб
+без
+будто
+человек
+чего
+раз
+тоже
+себе
+под
+жизнь
+будет
+ж
+тогда
+кто
+этот
+говорил
+того
+потому
+этого
+какой
+совсем
+ним
+здесь
+этом
+один
+почти
+мой
+тем
+чтобы
+нее
+кажется
+сейчас
+были
+куда
+зачем
+сказать
+всех
+никогда
+сегодня
+можно
+при
+наконец
+два
+об
+другой
+хоть
+после
+над
+больше
+тот
+через
+эти
+нас
+про
+всего
+них
+какая
+много
+разве
+сказала
+три
+эту
+моя
+впрочем
+хорошо
+свою
+этой
+перед
+иногда
+лучше
+чуть
+том
+нельзя
+такой
+им
+более
+всегда
+конечно
+всю
+между
+мной
+мною
+тебе
+тобой
+тобою
+нему
+эи
+ею
+нэи
+нею
+оно
+нам
+нами
+вами
+ими
+ними
+эта
+это
+эты
+этих
+этому
+этим
+этою
+этими
+та
+те
+ту
+той
+тех
+тому
+тою
+теми
+весь
+вся
+всей
+всему
+всем
+всею
+всеми
+сама
+само
+сами
+самого
+саму
+самих
+самой
+самому
+самим
+самою
+самими
+самом
