### 2016-04-14 f16d9b7
Add Appveyor CI
