### 2016-04-14 bd4e14a
Prepare for 0.1.2 release
