+++ NEWS.md
@@ -1,0 +1,13 @@
+# tokenizers 0.1.2
+
+- Add stopwords for several languages.
+- New stopword options to `tokenize_words()` and `tokenize_word_stems()`.
+
+# tokenizers 0.1.1
+
+- Fix failing test in non-UTF-8 locales.
+
+# tokenizers 0.1.0
+
+- Initial release with tokenizers for characters, words, word stems, sentences
+  paragraphs, n-grams, skip n-grams, lines, and regular expressions.
