+++ cran-comments.md
@@ -1,10 +1,5 @@
-## Response to CRAN maintainers' feedback
-
-CRAN maintainers pointed out that tests fail in an environment that uses the
-Latin-1 environment by default. I have specified that the files which are loaded
-for tests should be loaded as UTF-8. This should work for all Latin-1
-environments, since tests which previously were skipped on Win-builder now pass
-without being skipped.
+This is a minor update to add lists of stopwords and options to use them in the
+tokenization functions.
 
 ## Test environments
 
@@ -14,6 +9,4 @@ without being skipped.
 
 ## R CMD check results
 
-0 errors | 0 warnings | 1 note
-
-* The 1 NOTE pertains to a new release of the package.
+0 errors | 0 warnings | 0 notes
