### 2016-04-14 d240cdd
Fix CRAN comments
