### 2016-04-21 #33 @dselivanov
@kevinushey thank you for review. Probably I can say that I'm an author of `shingle_ngrams.cpp`. I'm not a professional c++ developer so find your comments related to c++ part very useful. Especially comments related to exporting `std` namespace and passing mutable objects by raw pointer.

One thing related to `RCPP_UNORDERED_MAP` and `RCPP_UNORDERED_SET`. This came from my experience with text2vec package. I noticed that `std::unordered_set` can't be compiled on windows - I checked it on win builder after several bug reports. This was with old gcc 4.6.3 toolchain, pretty sure it will work with new 4.9 toolchain. But interest thing is that there were no problems with `std::unordered_map`!
After googling a little bit I found post on Rcpp mailing list which suggests to use `RCPP_UNORDERED_SET` instead of `std::unordered_set`.
