### 2016-08-11 4a2ce76
Introductory vignette
