### 2016-08-11 023f801
Remove `using namespace std`
