### 2016-08-11 b16b146
Fix missing braces
