### 2016-08-11 b7be1d4
Scalar logical operators
