+++ R/utils.R
@@ -1,6 +1,6 @@
 simplify_list <- function(x, simplify) {
   stopifnot(is.logical(simplify))
-  if (simplify & length(x) == 1) x[[1]] else x
+  if (simplify && length(x) == 1) x[[1]] else x
 }
 
 check_input <- function(x) {
