### 2016-08-11 fa4b89e
Take advantage of testthat's `helper-data.R`
