+++ tests/testthat/test-utils.R
@@ -1,7 +1,5 @@
 context("Utils")
 
-source("data-for-tests.R")
-
 test_that("Inputs are verified correct", {
   expect_silent(check_input(letters))
   expect_silent(check_input(list(a = "a", b = "b")))
