### 2016-08-12 #33 @sckott
Thanks @lmullen - For those 2 outstanding issues, can you open issues for those in your repo, and we can consider these approved.
