### 2016-08-12 ropensci/tokenizers#20 @lmullen
Pass argument by reference using raw pointer
Issue from [rOpenSci onboarding](https://github.com/ropensci/onboarding/issues/33#issuecomment-239488688):

> https://github.com/lmullen/tokenizers/blob/d240cddbb1d91146b3a30b90f6fc25abd6919edf/src/shingle_ngrams.cpp#L24: This is mostly stylistic, but if you want to pass an argument by reference, and mutate it within that function, you should consider passing it using a raw pointer. This might seem strange, but the idea is that at the call site it's much more obvious that a function parameter might get mutated. (This is a habit I picked up from https://google.github.io/styleguide/cppguide.html#Reference_Arguments)
