### 2016-08-12 ropensci/tokenizers#21 @lmullen
Using long vectors
From [rOpenSci onboarding](https://github.com/ropensci/onboarding/issues/33#issuecomment-239488688):

> https://github.com/lmullen/tokenizers/blob/d240cddbb1d91146b3a30b90f6fc25abd6919edf/src/skip_ngrams.cpp#L9: If you want to support long vectors in R, you likely want to return and use R_xlen_t for indexing.
