### 2016-08-12 #33 @lmullen
@sckott: I've added them as issues.
