### 2016-08-12 #33 @lmullen
I transfered ownership to rOpenSci. Hope that's okay.
