### 2016-08-12 #33 @lmullen
It would be nice if more packages relied on tokenizers, so a little
publicity might help.

On Fri, Aug 12, 2016 at 1:02 PM, Scott Chamberlain <notifications@github.com
> wrote:

> sounds good, thanks, closing now, thanks again for submitting!
>
> would you be interested in doing a blog post on our blog about this?
>
