### 2016-08-12 #33 @sckott
true, would get more eyes on it, and hopefully uses in pkgs
